+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 1. [#5897](https://github.com/influxdata/chronograf/pull/5897): Upgrade golang to 1.18.
 1. [#5915](https://github.com/influxdata/chronograf/pull/5915): Upgrade github.com/lestrrat-go/jwx to v2.
 1. [#5933](https://github.com/influxdata/chronograf/pull/5933): Upgrade golang to 1.18.3 .
+1. [#5947](https://github.com/influxdata/chronograf/pull/5947): Use stable component keys.
 
 ## v1.9.4 [2022-03-22]